fix a bunch of merge conflicts
